Remove types module
