+++ bee-network/src/peer/error.rs
@@ -1,6 +1,8 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+#![cfg(feature = "full")]
+
 use libp2p::{Multiaddr, PeerId};
 
 /// [`PeerList`] errors.
