+++ bee-network/src/peer/mod.rs
@@ -1,8 +1,6 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-#![cfg(feature = "full")]
-
 pub mod error;
 pub mod list;
 pub mod meta;
